+++ drizzle/0007_dark_quasar.sql
@@ -1,0 +1,1 @@
+CREATE INDEX "user_library_episode_id_idx" ON "user_library" USING btree ("episode_id");